Use BASE_CXXFLAGS from xorg macros
The X.org macros package version 1.16.2 is the first to define
BASE_CXXFLAGS; however, we don't need to require version 1.16.2. If it's
available, then great. If not, $(BASE_CXXFLAGS) will be empty and
nothing terrible will happen.

Included in this commit is a fix for an error, now that we are using
BASE_CXXFLAGS, for shadowing variables from parent contexts.

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>