Add an xorg configuration file snippet to disable grabbing of test devices.
When a test creates a test trackpad device, an already running X server
will likely grab the event node. This causes the dummy X server to fail
to instantiate the device, and ends in test failure.

This will ensure that the normal X server does not attempt to grab the
test device.

Signed-off-by: Thomas Voß <thomas.voss@canonical.com>
Reviewed-by: Chase Douglas <chase.douglas@canonical.com>
Acked-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Chase Douglas <chase.douglas@canonical.com>