Provide a full server layout in dummy.conf
The X.org server uses the first server layout, and provides one if none
are found. If a layout is specified in /usr/share/X11/xorg.conf.d, the
dummy video device configuration may not be used.

This change ensures that the dummy video device is used by providing a
full server layout in the main configuration file.

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Reviewed-by: Stephen M. Webb <stephen.webb@bregmasoft.ca>