Ensure xorg-gtest implementation is included in xorg-gtest{-all, _main}.cpp
GNU automake puts -I. -I.. at the beginning of every invocation of g++.
This breaks xorg-gtest compilation if there is a file of the same name
as an xorg-gtest implementation file in ../src/ or ../../src. For
example, if compiling xorg-gtest-all.cpp in /home/user1/library/test
and there exists /home/user1/library/src/device.cpp, the library's
device.cpp file will be included instead of xorg-gtest's.

This change adds a -iquote option to xorg-gtest compilation performed in
Makefile-xorg-gtest.am. Directories specified with -iquote are searched
ahead of directories specified by -I.

This changes Makefile-xorg-gtest.am, which projects can copy to provide
rules for building xorg-gtest in autotools projects. For substantial
changes, projects will want to import the latest version of this file,
but this particular fix is only needed if a project runs into build
issues. If your project builds without issue, there is no need to
replace the existing version of Makefile-xorg-gtest.am.

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Acked-by: Peter Hutterer <peter.hutterer@who-t.net>