Disable man page documentation generation and installation
Now that we have marked some methods as deprecated, a "deprecated.3" man
page is created. Since the filename isn't namespaced, it could collide
with other documentation.

The man pages are kinda nice, but they really aren't that useful. It's
much better to use the HTML documentation. The simplest thing to do is
to remove the man page generation and installation.

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>