xserver: use a default timout for 1s for Terminate() and Kill()
Terminate/Kill will only wait for the server to exit for a nonzero timeout,
requiring callers to always specify a timeout for it to be useful.

The use-case of sending a SIGTERM to the server but not waiting for
it to shut down is narrow, so require those to actually specify a zero
timeout instead of everyone else.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Chase Douglas <chase.douglas@ubuntu.com>