Move default log/config setting back to environment
This was moved out of the Environment when the XServer class was added, but
it really doesn't belong there. Users of XServer should set custom config
and log file paths (if any) directly instead of having to undo the
libraries' defaults.

Move it back into the environment to avoid breaking users of that, but
really, this shouldn't even be in the environment either but rather in the caller
that needs those defaults.

Also move default path and display settings. Note that these are the same
defaults in the environment and in the server.

Since we can't rely on -displayfd yet, I'll leave the default display in
XServer to ensure the server always starts up (:133 is not likely to be in
use) and to auto-compose the display string for XOpenDisplay().

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>